storage: fix SR create/destroy tests scheduling conflict
SR create/destroy tests share the same disk as the package-scoped SR
fixtures used by other tests in the same directory. Depending on execution
order, those tests would fail if the SR fixture had already claimed the disk.

Change the SR fixture scope from 'package' to 'module' for all SR
types. Heavy infrastructure setup fixtures (package installation,
cluster configuration) remain package-scoped.

Move each TestXxxSRCreateDestroy class into a dedicated
test_create_destroy_xxx_sr.py file.
