+++ tests/packages/linux_image/conftest.py
@@ -10,7 +10,7 @@ from pkgfixtures import host_with_saved_yum_state
 
 from typing import Generator
 
-@pytest.fixture(scope="package")
+@pytest.fixture(scope="module")
 def host_with_perf(host_at_least_8_3: Host, host_with_saved_yum_state: Host) -> Generator[Host, None, None]:
     host = host_with_saved_yum_state
