+++ tests/packages/mlx/conftest.py
@@ -8,7 +8,7 @@ from pkgfixtures import host_with_saved_yum_state
 
 from typing import Generator
 
-@pytest.fixture(scope="package")
+@pytest.fixture(scope="module")
 def host_without_mlx_compat_loaded(host_with_saved_yum_state: Host) -> Generator[Host, None, None]:
     host = host_with_saved_yum_state
