+++ tests/storage/linstor/test_linstor_sr_intrapool_migration.py
@@ -18,7 +18,7 @@ from typing import Generator
 # And:
 # - access to XCP-ng RPM repository from hostA1
 
-@pytest.fixture(scope='package')
+@pytest.fixture(scope='module')
 def nfs_sr(host: Host, image_format: ImageFormat) -> Generator[SR, None, None]:
     """
     A NFS SR on first host.
