+++ tests/storage/xfs/conftest.py
@@ -17,7 +17,7 @@ from typing import Generator
 class XfsConfig:
     uninstall_xfs: bool = True
 
-@pytest.fixture(scope='package')
+@pytest.fixture(scope='module')
 def _xfs_config() -> XfsConfig:
     return XfsConfig()
 
@@ -25,7 +25,7 @@ def _xfs_config() -> XfsConfig:
 # To recreate host_with_xfsprogs for each image_format value, accept
 # image_format in the fixture arguments.
 # ref https://docs.pytest.org/en/7.1.x/how-to/fixtures.html#use-fixtures-in-classes-and-modules-with-usefixtures
-@pytest.fixture(scope='package')
+@pytest.fixture(scope='module')
 def host_with_xfsprogs(host: Host, image_format: ImageFormat, _xfs_config: XfsConfig) -> Generator[Host, None, None]:
     assert not host.file_exists('/usr/sbin/mkfs.xfs'), \
         "xfsprogs must not be installed on the host at the beginning of the tests"
@@ -36,7 +36,7 @@ def host_with_xfsprogs(host: Host, image_format: ImageFormat, _xfs_config: XfsCo
     if _xfs_config.uninstall_xfs:
         host.yum_restore_saved_state()
 
-@pytest.fixture(scope='package')
+@pytest.fixture(scope='module')
 def xfs_sr(
     unused_512B_disks: dict[Host, list[Host.BlockDeviceInfo]],
     host_with_xfsprogs: Host,
