+++ tests/storage/zfsvol/test_create_destroy_zfsvol_sr.py
@@ -1,0 +1,30 @@
+from __future__ import annotations
+
+import pytest
+
+from lib.common import vm_image
+from lib.host import Host
+
+# Requirements:
+# - one XCP-ng host >= 8.3 with an additional unused disk for the SR
+# - access to XCP-ng RPM repository from the host
+
+pytestmark = pytest.mark.usefixtures("host_at_least_8_3")
+
+
+class TestZfsvolSRCreateDestroy:
+    """
+    Tests that do not use fixtures that setup the SR or import VMs,
+    because they precisely need to test SR creation and destruction,
+    and VM import.
+    """
+
+    def test_create_and_destroy_sr(self, sr_disk_wiped: str, host_with_zfsvol: Host) -> None:
+        host = host_with_zfsvol
+        # Create and destroy tested in the same test to leave the host as unchanged as possible
+        sr = host.sr_create('zfs-vol', "ZFS-local-SR-test", {'device': '/dev/' + sr_disk_wiped}, verify=True)
+        # import a VM in order to detect vm import issues here rather than in the vm_on_zfsvol_sr fixture used in
+        # the next tests, because errors in fixtures break teardown
+        vm = host.import_vm(vm_image('mini-linux-x86_64-bios'), sr_uuid=sr.uuid)
+        vm.destroy(verify=True)
+        sr.destroy(verify=True)
